Move Camera_Model abstract class into camera_models
